# Conflicts:
#	unimall-admin/config/dev.env.js

+++ unimall-data/src/main/java/com/iotechn/unimall/data/dto/goods/GroupShopSkuDTO.java
@@ -14,7 +14,7 @@ public class GroupShopSkuDTO extends SpuDTO {
 
     private Integer skuGroupShopPrice;
 
-    private Integer groupShopId;
+    private Long groupShopId;
 
     private Long skuId;
